doc: add code document + some refactor
