+++ .gitignore
@@ -23,3 +23,6 @@ go.work.sum
 
 # env file
 .env
+
+client
+client.log
