+++ go.mod
@@ -1,15 +1,11 @@
-module github.com/dohaelsawy/codescalers/datetime-client
+module github.com/codescalersinternships/Datetime-HTTP-Client-Doha
 
 go 1.21.0
 
 require (
+	github.com/cenkalti/backoff/v4 v4.3.0
 	github.com/joho/godotenv v1.5.1
 	github.com/sirupsen/logrus v1.9.3
 )
 
-require github.com/stretchr/testify v1.9.0 // indirect
-
-require (
-	github.com/cenkalti/backoff/v4 v4.3.0
-	golang.org/x/sys v0.0.0-20220715151400-c0bba94af5f8 // indirect
-)
+require golang.org/x/sys v0.0.0-20220715151400-c0bba94af5f8 // indirect
